rename compiler_support folder to better name
