+++ include/ccmath/internal/predef/compiler_suppression/gcc_compiler_suppression.hpp
