+++ include/ccmath/internal/support/math_support.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "ccmath/internal/predef/compiler_support/msvc_compiler_suppression.hpp"
+#include "ccmath/internal/predef/compiler_suppression/msvc_compiler_suppression.hpp"
 #include "ccmath/internal/predef/has_builtin.hpp"
 #include "ccmath/internal/support/is_constant_evaluated.hpp"
 #include "ccmath/internal/support/type_traits.hpp"
